Migrate the `wasm/lists.rs` test to the new framework
