--- crates/test-rust-wasm/src/bin/lists.rs
@@ -1,3 +1,0 @@
-include!("../../../../tests/runtime/lists/wasm.rs");
-
-fn main() {}
